pyproject.toml
[project]
name = "expanse"
version = "0.1.0"
description = "Tactical space combat sim inspired by The Expanse."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.7",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/expanse"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
